pyproject.toml
[project]
name = "autowiki"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "arq>=0.26",
    "redis>=5.0",
    "sqlalchemy>=2.0",
    "aiosqlite>=0.20",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "pyyaml>=6.0",
    "gitpython>=3.1",
    "tree-sitter>=0.23",
    "tree-sitter-python>=0.23",
    "tree-sitter-javascript>=0.23",
    "tree-sitter-typescript>=0.23",
    "tree-sitter-java>=0.23",
    "tree-sitter-go>=0.21",
    "tree-sitter-rust>=0.23",
    "tree-sitter-c>=0.21",
    "tree-sitter-cpp>=0.23",
    "tree-sitter-c-sharp>=0.23",
    "tree-sitter-kotlin>=0.23",
    "langchain-text-splitters>=0.3",
    "faiss-cpu>=1.8",
    "anthropic>=0.40",
    "openai>=1.50",
    "google-genai>=0.2",
    "typer>=0.12",
    "httpx>=0.27",
    "websockets>=13.0",
]

[project.scripts]
autowiki = "cli.main:app"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "httpx>=0.27",
    "pytest-cov>=5.0",
    "playwright>=1.48",
    "pytest-playwright>=0.5",
    "ruff>=0.11",
]

[tool.hatch.build.targets.wheel]
packages = ["api", "worker", "shared", "cli"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"] # Enable basic error checking, unreferenced variables checking, import sorting (isort), and syntax upgrade checking
